// BlazorAutoApp.React/app/api/generated/schema.d.ts
export interface components {
    schemas: {
        AuthorBookListItemResponse: {
            /** Format: int32 */
            id?: number | string;
            seedKey: string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
        BookListItemResponse: {
            /** Format: int32 */
            id?: number | string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
        CreateBookRequest: {
            title: string;
            author?: null | string;
            url?: null | string;
        };
        CreateBookResponse: {
            /** Format: int32 */
            id?: number | string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
        GetAuthorBookResponse: {
            /** Format: int32 */
            id?: number | string;
            seedKey: string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
        GetAuthorBooksResponse: {
            books: components["schemas"]["AuthorBookListItemResponse"][];
        };
        GetBookResponse: {
            /** Format: int32 */
            id?: number | string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
        GetBooksResponse: {
            books: components["schemas"]["BookListItemResponse"][];
        };
        HttpValidationProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number | string;
            detail?: null | string;
            instance?: null | string;
            errors?: {
                [key: string]: string[];
            };
        };
        ProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number | string;
            detail?: null | string;
            instance?: null | string;
        };
        UpdateBookRequest: {
            /** Format: int32 */
            id?: number | string;
            title: string;
            author?: null | string;
            url?: null | string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}